refactor: move remaining init hooks to linter modules Signed-off-by: Gregor Zeitlinger <gregor.zeitlinger@grafana.com>
