--- src/init/hooks.rs
@@ -1,4 +1,0 @@
-pub(crate) mod rumdl;
-pub(crate) mod rustfmt;
-pub(crate) mod taplo;
-pub(crate) mod yamllint;
