+++ src/linters/mod.rs
@@ -4,7 +4,10 @@ pub mod flint_setup;
 pub mod license_header;
 pub mod lychee;
 pub mod renovate_deps;
+pub mod rumdl;
+pub mod rustfmt;
 pub mod taplo;
+pub mod yamllint;
 
 /// Build a [`tokio::process::Command`] for the given argv.
 ///
