+++ src/linters/rustfmt.rs
@@ -3,7 +3,7 @@ use std::path::Path;
 
 use crate::registry::InitHookContext;
 
-pub(crate) fn run(ctx: &dyn InitHookContext) -> Result<bool> {
+pub(crate) fn init(ctx: &dyn InitHookContext) -> Result<bool> {
     generate_config(ctx.config_dir(), ctx.line_length())
 }
